# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.AtRiskStudent:
    properties:
      attendance_percentage:
        type: number
      current_streak:
        type: integer
      days_inactive:
        type: integer
      final_score:
        type: number
      reasons:
        items:
          type: string
        type: array
      student_id:
        type: string
      student_name:
        type: string
    type: object
...
  models.MaterialSummary:
    properties:
      enable_downloads:
        type: boolean
      file_url:
        type: string
      is_active:
        type: boolean
      is_prerequisite:
        type: boolean
      material_name:
        type: string
      material_type:
        type: string
      short_id:
        type: string
    type: object
...
  models.ModuleScheduleEntry:
    properties:
      module_name:
        type: string
      module_short_id:
        type: string
      release_date:
        type: string
    type: object
...
  models.ModuleWithSections:
    properties:
      is_active:
        type: boolean
      is_released:
        description: |-
          IsReleased/ReleaseDate are only populated by the batch-scoped curriculum
          endpoint (GET /batches/{short_id}/curriculum) — nil when this struct is
          returned by the plain course-level curriculum endpoint, which has no
          batch context to schedule against.
        type: boolean
      max_view_duration:
        type: string
      module_branch:
        type: string
      module_name:
        type: string
      order_index:
        type: integer
      release_date:
        type: string
      sections:
        items:
          $ref: '#/definitions/models.SectionSummary'
        type: array
      short_id:
        type: string
      watch_time_minutes:
        type: integer
    type: object
...
  models.SectionSummary:
    properties:
      is_active:
        type: boolean
      is_prerequisite:
        type: boolean
      materials:
        items:
          $ref: '#/definitions/models.MaterialSummary'
        type: array
      section_name:
        type: string
      short_description:
        type: string
      short_id:
        type: string
    type: object
...
  models.SetModuleScheduleInput:
    properties:
      release_date:
        example: "2026-08-01"
        type: string
    required:
    - release_date
    type: object
...
  models.StudentStreak:
    properties:
      current_streak:
        type: integer
      last_active_date:
        type: string
      longest_streak:
        type: integer
      student_id:
        type: string
      student_name:
        type: string
      total_active_days:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: umangagarwal9099@gmail.com
    name: Umang Agarwal
  description: 'Vedex backend API. Authenticate via POST /api/v1/auth/login to get
    a JWT, then pass it as `Authorization: Bearer <token>` on protected endpoints.'
  title: Vedex API
  version: "1.0"
paths:
...
  /batches/{short_id}/at-risk:
    get:
      description: Flags students with low attendance (<60%), a low final score (<40%,
        only once they have graded work), or no activity in 7+ days (only once the
        batch has been running that long). Only returns students who tripped at least
        one flag.
      parameters:
      - description: Batch short ID
        in: path
        name: short_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.AtRiskStudent'
            type: array
        "404":
          description: Batch not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a batch's at-risk students
      tags:
      - engagement
...
  /batches/{short_id}/curriculum:
    get:
      description: Returns the batch's course curriculum, annotating every module
        with is_released and (if scheduled) release_date. A module with no configured
        schedule is always released. Students don't see sections/materials for a module
        that isn't released yet.
      parameters:
      - description: Batch short ID
        in: path
        name: short_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ModuleWithSections'
            type: array
        "404":
          description: Batch not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a batch's curriculum (with module release status)
      tags:
      - curriculum
...
  /batches/{short_id}/modules/{module_short_id}/schedule:
    delete:
      description: Removes a module's release schedule for this batch, reverting it
        to "released from day one". Restricted to super_admin / team_lead / mentor.
      parameters:
      - description: Batch short ID
        in: path
        name: short_id
        required: true
        type: string
      - description: Module short ID
        in: path
        name: module_short_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Batch, module, or schedule not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Unschedule a module's release for a batch
      tags:
      - curriculum
    patch:
      consumes:
      - application/json
      description: Sets (or updates) the date a module becomes visible to this batch's
        students. Restricted to super_admin / team_lead / mentor.
      parameters:
      - description: Batch short ID
        in: path
        name: short_id
        required: true
        type: string
      - description: Module short ID
        in: path
        name: module_short_id
        required: true
        type: string
      - description: Release date
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.SetModuleScheduleInput'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Validation error
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Batch or module not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Schedule a module's release for a batch
      tags:
      - curriculum
  /batches/{short_id}/modules/schedule:
    get:
      description: Returns every module that has a configured release date for this
        batch.
      parameters:
      - description: Batch short ID
        in: path
        name: short_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ModuleScheduleEntry'
            type: array
        "404":
          description: Batch not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List a batch's module release schedule
      tags:
      - curriculum
...
  /users/{id}/streak:
    get:
      description: '"Active" means the student submitted an assignment/project, attempted
        an exam, or attended a session on that calendar day.'
      parameters:
      - description: Student user ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StudentStreak'
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a student's activity streak
      tags:
      - engagement
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Format: `Bearer <token>`'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"